requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
dropbox==12.0.2
fastapi==0.115.11
fonttools==4.56.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
ply==3.11
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.0.1
python-multipart==0.0.20
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
stone==3.3.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
redis==5.0.1
rq==1.15.1
apify-client==1.7.0
gspread==6.0.0
google-auth==2.27.0
pandas==2.2.0
openpyxl==3.1.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
boto3>=1.28.0
botocore>=1.31.0
pypdf>=4.0.0
